+++ src/models/RecipesList.ts
@@ -1,3 +1,4 @@
+import { arrayFilter } from '@noeldemartin/utils';
 import type { Relation } from 'soukai';
 import type { RouteLocationRaw } from 'vue-router';
 import type { SolidBelongsToManyRelation } from 'soukai-solid';
@@ -5,8 +6,9 @@ import type { SolidBelongsToManyRelation } from 'soukai-solid';
 import Model from './RecipesList.schema';
 import RecipesListItem from './RecipesListItem';
 import type Recipe from './Recipe';
+import type RecipesCollection from './contracts/RecipesCollection';
 
-export default class RecipesList extends Model {
+export default class RecipesList extends Model implements RecipesCollection {
 
     public isRegistered = false;
     declare public items?: RecipesListItem[];
@@ -15,6 +17,14 @@ export default class RecipesList extends Model {
         RecipesListItem,
         typeof RecipesListItem
     >;
+
+    public get recipes(): Recipe[] | undefined {
+        if (!this.isRelationLoaded('items')) {
+            return;
+        }
+
+        return arrayFilter(this.items?.map(({ recipe }) => recipe) ?? []);
+    }
 
     public itemsRelationship(): Relation {
         return this