+++ src/routing/pages/viewer/components/ViewerRecipe.vue
@@ -1,9 +1,9 @@
 <template>
     <RecipePage
-        v-if="$viewer.recipe"
+        v-if="recipe"
         v-element-transitions="{
             name: 'viewer-recipe',
-            id: $viewer.recipe.url,
+            id: recipe.url,
             transitions: {
                 enter: enterTransition,
                 leave: {
@@ -13,23 +13,23 @@
                 'recipe-card': transitionToCard,
             },
         }"
-        :recipe="$viewer.recipe"
+        :recipe="recipe"
     >
         <template #title>
             <h1
                 id="viewer-recipe-title"
                 class="viewer-recipe--header-title text-4xl font-semibold text-white text-shadow"
             >
-                <span class="viewer-recipe--header-title-text">{{ $viewer.recipe.name }}</span>
+                <span class="viewer-recipe--header-title-text">{{ recipe.name }}</span>
             </h1>
         </template>
 
         <template #actions>
             <CoreButton
-                v-if="$viewer.recipeInCookbook"
+                v-if="recipeInCookbook"
                 secondary
                 route="recipes.show"
-                :route-params="{ recipe: $viewer.recipeInCookbook?.slug }"
+                :route-params="{ recipe: recipeInCookbook?.slug }"
             >
                 {{ $t('viewer.recipe.inCookbook') }}
             </CoreButton>
@@ -40,16 +40,16 @@
             </CoreButton>
 
             <ViewerRecipeCreator
-                v-if="list"
+                v-if="collection"
                 :prefix="$t('viewer.recipe.creatorPrefix')"
-                :list="list"
+                :collection="collection"
             />
 
             <CoreLink
-                v-if="list"
+                v-if="collection"
                 class="text-sm"
                 route="viewer"
-                :route-query="{ url: list.url }"
+                :route-query="{ url: collection.url }"
             >
                 {{ $t('viewer.recipe.viewAll') }}
             </CoreLink>
@@ -67,12 +67,25 @@ import UI from '@/framework/core/facades/UI';
 import { translate } from '@/framework/utils/translate';
 import { watchDebouncedEffect } from '@/framework/utils/vue';
 
+import Cookbook from '@/services/facades/Cookbook';
 import Recipe from '@/models/Recipe';
 import Viewer from '@/services/facades/Viewer';
 
 import { enterTransition, transitionToCard } from './ViewerRecipe.transitions';
 
-const list = $computed(() => Viewer.recipe?.lists?.[0]);
+const recipe = $computed(() => Viewer.recipe);
+const list = $computed(() => recipe && recipe?.lists?.[0]);
+const container = $computed(() => recipe && Viewer.getRecipeContainer(recipe));
+const collection = $computed(() => list ?? container);
+const recipeInCookbook = $computed(() => {
+    if (!recipe?.url) {
+        return null;
+    }
+
+    return Cookbook.recipes.find(
+        localRecipe => localRecipe.url === recipe.url || localRecipe.externalUrls.includes(recipe.url),
+    ) ?? null;
+});
 
 function showUnsupportedIndexedDBAlert() {
     const message = translate('errors.unsupportedBrowser_indexedDBFeature');
@@ -88,7 +101,7 @@ function showUnsupportedIndexedDBAlert() {
 }
 
 async function importRecipe(): Promise<void> {
-    if (!Viewer.recipe) {
+    if (!recipe) {
         return;
     }
 
@@ -98,15 +111,27 @@ async function importRecipe(): Promise<void> {
         return;
     }
 
-    const remoteRecipe = tap(Viewer.recipe.clone(), recipe => recipe.unloadRelation('lists'));
-    const recipe = await Recipe.newFromJsonLD(remoteRecipe.toExternalJsonLD({ includeIds: false }));
+    const remoteRecipe = tap(recipe.clone(), recipe => recipe.unloadRelation('lists'));
+    const newRecipe = await Recipe.newFromJsonLD(remoteRecipe.toExternalJsonLD({ includeIds: false }));
 
-    recipe.externalUrls = arrayUnique([...recipe.externalUrls, remoteRecipe.url]);
+    newRecipe.externalUrls = arrayUnique([...newRecipe.externalUrls, remoteRecipe.url]);
 
-    await recipe.save();
+    await newRecipe.save();
     await Router.push({ name: 'home' });
-    await Cloud.syncIfOnline(recipe);
+    await Cloud.syncIfOnline(newRecipe);
 }
 
-watchDebouncedEffect($$(list), 300, () => list && Viewer.preload(list.url));
+watchDebouncedEffect($$(collection), 300, () => {
+    if (list) {
+        Viewer.preloadList(list);
+
+        return;
+    }
+
+    if (recipe) {
+        Viewer.preloadContainer(recipe.requireContainerUrl());
+
+        return;
+    }
+});
 </script>